# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Swift Tow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  user: "user",
  driver: "driver",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** Ghana phone number starting with +233 */
  phone: string;
  role?: RegisterRequestRole;
}

export type LoginRequestRole =
  (typeof LoginRequestRole)[keyof typeof LoginRequestRole];

export const LoginRequestRole = {
  user: "user",
  driver: "driver",
} as const;

export interface LoginRequest {
  email: string;
  password: string;
  role?: LoginRequestRole;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  driver: "driver",
  admin: "admin",
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  phone: string;
  role: UserRole;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateProfileRequest {
  userId: string;
  name?: string;
  phone?: string;
  avatarUrl?: string | null;
}

export interface Coordinates {
  latitude: number;
  longitude: number;
}

export type NewTowRequestTowType =
  (typeof NewTowRequestTowType)[keyof typeof NewTowRequestTowType];

export const NewTowRequestTowType = {
  flatbed: "flatbed",
  hook_chain: "hook_chain",
  repair: "repair",
} as const;

export interface NewTowRequest {
  userId: string;
  userName: string;
  userPhone: string;
  towType: NewTowRequestTowType;
  pickupLocation: Coordinates;
  pickupAddress: string;
  dropoffAddress?: string | null;
  /** e.g. Toyota Camry - GT 442-22 */
  vehicleDetails: string;
}

export type TowRequestUpdateStatus =
  (typeof TowRequestUpdateStatus)[keyof typeof TowRequestUpdateStatus];

export const TowRequestUpdateStatus = {
  pending: "pending",
  accepted: "accepted",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface TowRequestUpdate {
  status: TowRequestUpdateStatus;
  driverId?: string | null;
  driverLocation?: Coordinates;
  /** Estimated arrival in minutes */
  estimatedArrival?: number;
}

export type TowRequestTowType =
  (typeof TowRequestTowType)[keyof typeof TowRequestTowType];

export const TowRequestTowType = {
  flatbed: "flatbed",
  hook_chain: "hook_chain",
  repair: "repair",
} as const;

export type TowRequestStatus =
  (typeof TowRequestStatus)[keyof typeof TowRequestStatus];

export const TowRequestStatus = {
  pending: "pending",
  accepted: "accepted",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface TowRequest {
  id: string;
  userId: string;
  userName: string;
  userPhone: string;
  towType: TowRequestTowType;
  status: TowRequestStatus;
  pickupLocation: Coordinates;
  pickupAddress: string;
  dropoffAddress?: string | null;
  vehicleDetails: string;
  driverId?: string | null;
  estimatedArrival?: number | null;
  amount?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface Driver {
  id: string;
  name: string;
  email: string;
  phone: string;
  isOnline: boolean;
  currentLocation?: Coordinates;
  avatarUrl?: string | null;
  rating: number;
  totalTrips: number;
  activeJobId?: string | null;
}

export interface DriverStatusUpdate {
  isOnline: boolean;
  currentLocation?: Coordinates;
}

export type TripTowType = (typeof TripTowType)[keyof typeof TripTowType];

export const TripTowType = {
  flatbed: "flatbed",
  hook_chain: "hook_chain",
  repair: "repair",
} as const;

export type TripPaymentMethod =
  | (typeof TripPaymentMethod)[keyof typeof TripPaymentMethod]
  | null;

export const TripPaymentMethod = {
  mtn_momo: "mtn_momo",
  telecel_cash: "telecel_cash",
  at_money: "at_money",
  cash: "cash",
} as const;

export type TripPaymentStatus =
  (typeof TripPaymentStatus)[keyof typeof TripPaymentStatus];

export const TripPaymentStatus = {
  pending: "pending",
  paid: "paid",
} as const;

export interface Trip {
  id: string;
  towRequestId: string;
  userId: string;
  driverId: string;
  driverName: string;
  pickupAddress: string;
  dropoffAddress?: string | null;
  vehicleDetails: string;
  towType: TripTowType;
  amount: number;
  paymentMethod?: TripPaymentMethod;
  paymentStatus: TripPaymentStatus;
  completedAt: string;
}

export type ProcessPaymentRequestMethod =
  (typeof ProcessPaymentRequestMethod)[keyof typeof ProcessPaymentRequestMethod];

export const ProcessPaymentRequestMethod = {
  mtn_momo: "mtn_momo",
  telecel_cash: "telecel_cash",
  at_money: "at_money",
  cash: "cash",
} as const;

export interface ProcessPaymentRequest {
  tripId: string;
  amount: number;
  method: ProcessPaymentRequestMethod;
  phoneNumber?: string | null;
}

export interface PaymentResult {
  success: boolean;
  transactionId?: string | null;
  message: string;
}

export interface DashboardStats {
  onlineDrivers: number;
  activeJobs: number;
  totalEarnings: number;
  totalTrips: number;
  pendingRequests: number;
  recentRequests: TowRequest[];
}

export type GetProfileParams = {
  userId: string;
};

export type ListTowRequestsParams = {
  userId?: string;
  driverId?: string;
  status?: ListTowRequestsStatus;
};

export type ListTowRequestsStatus =
  (typeof ListTowRequestsStatus)[keyof typeof ListTowRequestsStatus];

export const ListTowRequestsStatus = {
  pending: "pending",
  accepted: "accepted",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ListDriversParams = {
  online?: boolean;
};

export type ListTripsParams = {
  userId: string;
};