# lib/api-zod/src/generated/types/driverStatusUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Swift Tow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Coordinates } from "./coordinates";

export interface DriverStatusUpdate {
  isOnline: boolean;
  currentLocation?: Coordinates;
}