# lib/api-zod/src/generated/types/newTowRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Swift Tow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Coordinates } from "./coordinates";
import type { NewTowRequestTowType } from "./newTowRequestTowType";

export interface NewTowRequest {
  userId: string;
  userName: string;
  userPhone: string;
  towType: NewTowRequestTowType;
  pickupLocation: Coordinates;
  pickupAddress: string;
  dropoffAddress?: string | null;
  /** e.g. Toyota Camry - GT 442-22 */
  vehicleDetails: string;
}